backend/requirements.txt
fastapi
uvicorn
sqlmodel
scikit-learn
pandas
requests
python-multipart
passlib[bcrypt]
python-dotenv
